pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "arxiv>=2.2.0",
    "diffusers>=0.35.1",
    "ipykernel>=6.30.1",
    "langchain-community>=0.3.30",
    "langchain-experimental>=0.3.4",
    "langchain-gemini>=0.1.1",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.34",
    "matplotlib>=3.10.6",
    "ollama>=0.6.0",
    "pandas>=2.3.2",
    "pymupdf>=1.26.4",
    "pyperclip>=1.11.0",
    "pytest>=8.4.2",
    "rich>=14.1.0",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "tensorflow>=2.20.0",
    "tf-keras>=2.20.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
    "wikipedia>=1.4.0",
]

# [tool.uv.sources]
# torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
